refactor: Move/remove files for testing
