+++ public/collision/index.js
@@ -1,0 +1,3 @@
+export const PlayerOverheadObstacleHeight = {
+	STANDING: 73,
+};
