+++ public/collision/textures/textures.json
@@ -1,15 +1,8 @@
 [
 	"debug.jpg",
-	"metal/metalcombine001.jpg",
-	"metal/metalcombine001_normal.jpg",
 	"normal.jpg",
-	"plaster/plasterceiling002b.jpg",
-	"plaster/plasterceiling002b_normal.jpg",
-	"plaster/plasterwall030a.jpg",
 	"plaster/plasterwall030c.jpg",
 	"plaster/plasterwall030c_normal.jpg",
-	"tile/tilefloor018a.jpg",
-	"tile/tilefloor018a_normal.jpg",
 	"wood/wooddoor048a_normal.jpg",
 	"wood/wooddoor048a.jpg"
 ]
