+++ src/Loader/index.js
@@ -2,3 +2,4 @@ export {CaptureSessionLoader} from "./CaptureSessionLoader.js";
 export {ImageBitmapLoader} from "./ImageBitmapLoader.js";
 export {Loader} from "./Loader.js";
 export {ShaderLoader} from "./ShaderLoader.js";
+export {SSDLoader} from "./SSDLoader.js";
